@@ -83,14 +83,14 @@
     "packaging": "jar"
   },
   "dependencies": {
-    "@angular/common": "17.3.6",
-    "@angular/compiler": "17.3.6",
-    "@angular/core": "17.3.6",
-    "@angular/forms": "17.3.6",
-    "@angular/localize": "17.3.6",
-    "@angular/platform-browser": "17.3.6",
-    "@angular/platform-browser-dynamic": "17.3.6",
-    "@angular/router": "17.3.6",
+    "@angular/common": "17.3.9",
+    "@angular/compiler": "17.3.9",
+    "@angular/core": "17.3.9",
+    "@angular/forms": "17.3.9",
+    "@angular/localize": "17.3.9",
+    "@angular/platform-browser": "17.3.9",
+    "@angular/platform-browser-dynamic": "17.3.9",
+    "@angular/router": "17.3.9",
     "@fortawesome/angular-fontawesome": "0.14.1",
     "@fortawesome/fontawesome-svg-core": "6.5.2",
     "@fortawesome/free-solid-svg-icons": "6.5.2",
@@ -103,49 +103,49 @@
     "ngx-infinite-scroll": "17.0.0",
     "rxjs": "7.8.1",
     "tslib": "2.6.2",
-    "zone.js": "0.14.4"
+    "zone.js": "0.14.6"
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "17.0.2",
     "@angular-builders/jest": "17.0.3",
-    "@angular-devkit/build-angular": "17.3.6",
-    "@angular-eslint/eslint-plugin": "17.3.0",
-    "@angular/cli": "17.3.6",
-    "@angular/compiler-cli": "17.3.6",
-    "@angular/service-worker": "17.3.6",
+    "@angular-devkit/build-angular": "17.3.7",
+    "@angular-eslint/eslint-plugin": "17.5.2",
+    "@angular/cli": "17.3.7",
+    "@angular/compiler-cli": "17.3.9",
+    "@angular/service-worker": "17.3.9",
     "@types/jest": "29.5.12",
     "@types/node": "20.11.25",
-    "@typescript-eslint/eslint-plugin": "7.7.1",
-    "@typescript-eslint/parser": "7.7.1",
+    "@typescript-eslint/eslint-plugin": "7.11.0",
+    "@typescript-eslint/parser": "7.11.0",
     "browser-sync": "3.0.2",
     "browser-sync-webpack-plugin": "2.3.0",
     "buffer": "6.0.3",
     "concurrently": "8.2.2",
     "copy-webpack-plugin": "12.0.2",
-    "cypress": "13.8.1",
+    "cypress": "13.10.0",
     "cypress-audit": "1.1.0",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-cypress": "3.0.2",
-    "eslint-webpack-plugin": "4.1.0",
+    "eslint-plugin-cypress": "3.3.0",
+    "eslint-webpack-plugin": "4.2.0",
     "folder-hash": "4.0.4",
-    "generator-jhipster": "8.4.0",
+    "generator-jhipster": "8.5.0",
     "husky": "9.0.11",
     "jest": "29.7.0",
     "jest-date-mock": "1.0.10",
     "jest-environment-jsdom": "29.7.0",
     "jest-junit": "16.0.0",
-    "jest-preset-angular": "14.0.3",
+    "jest-preset-angular": "14.1.0",
     "jest-sonar": "0.2.16",
     "lighthouse": "12.0.0",
-    "lint-staged": "15.2.2",
+    "lint-staged": "15.2.5",
     "merge-jsons-webpack-plugin": "2.0.1",
     "prettier": "3.2.5",
     "prettier-plugin-java": "2.6.0",
     "prettier-plugin-packagejson": "2.5.0",
-    "rimraf": "5.0.5",
-    "swagger-ui-dist": "5.17.2",
-    "ts-jest": "29.1.2",
+    "rimraf": "5.0.7",
+    "swagger-ui-dist": "5.17.14",
+    "ts-jest": "29.1.4",
     "typescript": "5.4.5",
     "wait-on": "7.2.0",
     "webpack-bundle-analyzer": "4.10.2",
@@ -153,7 +153,7 @@
     "webpack-notifier": "1.15.0"
   },
   "engines": {
-    "node": ">=20.12.2"
+    "node": ">=20.14.0"
   },
   "cacheDirectories": [
     "node_modules"
