@@ -1,4 +1,4 @@
-import { Directive, OnInit, ElementRef, Renderer2, inject, Input } from '@angular/core';
+import { Directive, OnInit, ElementRef, Renderer2, inject, input } from '@angular/core';
 import { TranslateService, LangChangeEvent } from '@ngx-translate/core';
 
 @Directive({
@@ -6,7 +6,7 @@ import { TranslateService, LangChangeEvent } from '@ngx-translate/core';
   selector: '[jhiActiveMenu]',
 })
 export default class ActiveMenuDirective implements OnInit {
-  @Input() jhiActiveMenu?: string;
+  jhiActiveMenu = input();
 
   private el = inject(ElementRef);
   private renderer = inject(Renderer2);
@@ -21,7 +21,7 @@ export default class ActiveMenuDirective implements OnInit {
   }
 
   updateActiveFlag(selectedLanguage: string): void {
-    if (this.jhiActiveMenu === selectedLanguage) {
+    if (this.jhiActiveMenu() === selectedLanguage) {
       this.renderer.addClass(this.el.nativeElement, 'active');
     } else {
       this.renderer.removeClass(this.el.nativeElement, 'active');
