@@ -6,7 +6,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Generated(value = "JHipster", comments = "Generated by JHipster 8.4.0")
+@Generated(value = "JHipster", comments = "Generated by JHipster 8.5.0")
 @Retention(RetentionPolicy.SOURCE)
 @Target({ ElementType.TYPE })
 public @interface GeneratedByJHipster {
