@@ -1,18 +1,18 @@
 import { IUser } from './user.model';
 
 export const sampleWithRequiredData: IUser = {
-  id: 7665,
-  login: 'XsgD',
+  id: 9396,
+  login: 'qj4pT8@fThe2\\qHnO\\YH7tyC',
 };
 
 export const sampleWithPartialData: IUser = {
-  id: 32722,
-  login: '2`{@7f\\&Wq\\4gbOs\\CwDh\\&mr\\5-n',
+  id: 1979,
+  login: 'm8_OM',
 };
 
 export const sampleWithFullData: IUser = {
-  id: 30499,
-  login: '-PIP',
+  id: 30730,
+  login: 'SZ2@6fZn\\]Ggso',
 };
 Object.freeze(sampleWithRequiredData);
 Object.freeze(sampleWithPartialData);
