@@ -1,7 +1,7 @@
 jest.mock('app/core/auth/account.service');
 jest.mock('app/login/login.service');
 
-import { ElementRef } from '@angular/core';
+import { ElementRef, signal } from '@angular/core';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { FormBuilder } from '@angular/forms';
 import { Router, Navigation } from '@angular/router';
@@ -90,7 +90,7 @@ describe('LoginComponent', () => {
       const node = {
         focus: jest.fn(),
       };
-      comp.username = new ElementRef(node);
+      comp.username = signal<ElementRef>(new ElementRef(node));
 
       // WHEN
       comp.ngAfterViewInit();
