@@ -1,7 +1,9 @@
 jest.mock('app/core/auth/account.service');
 
-import { Component, ElementRef, ViewChild, WritableSignal, signal } from '@angular/core';
+import { Component, ElementRef, WritableSignal, signal, viewChild } from '@angular/core';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TestBed, waitForAsync } from '@angular/core/testing';
+import { TranslateModule } from '@ngx-translate/core';
 
 import { AccountService } from 'app/core/auth/account.service';
 import { Account } from 'app/core/auth/account.model';
@@ -12,8 +14,7 @@ import HasAnyAuthorityDirective from './has-any-authority.directive';
   template: ` <div *jhiHasAnyAuthority="'ROLE_ADMIN'" #content></div> `,
 })
 class TestHasAnyAuthorityDirectiveComponent {
-  @ViewChild('content', { static: false })
-  content?: ElementRef;
+  content = viewChild<ElementRef>('content');
 }
 
 describe('HasAnyAuthorityDirective tests', () => {
@@ -22,7 +23,7 @@ describe('HasAnyAuthorityDirective tests', () => {
 
   beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
-      imports: [HasAnyAuthorityDirective],
+      imports: [HasAnyAuthorityDirective, HttpClientTestingModule, TranslateModule.forRoot()],
       declarations: [TestHasAnyAuthorityDirectiveComponent],
       providers: [AccountService],
     });
@@ -60,7 +61,7 @@ describe('HasAnyAuthorityDirective tests', () => {
       fixture.detectChanges();
 
       // THEN
-      expect(comp.content).toBeUndefined();
+      expect(comp.content()).toBeUndefined();
     });
   });
 
@@ -76,7 +77,7 @@ describe('HasAnyAuthorityDirective tests', () => {
       fixture.detectChanges();
 
       // THEN
-      expect(comp.content).toBeDefined();
+      expect(comp.content()).toBeDefined();
 
       // GIVEN
       currentAccount.set(null);
@@ -85,7 +86,7 @@ describe('HasAnyAuthorityDirective tests', () => {
       fixture.detectChanges();
 
       // THEN
-      expect(comp.content).toBeUndefined();
+      expect(comp.content()).toBeUndefined();
 
       // WHEN
       currentAccount.set({ activated: true, authorities: ['foo'] } as any);
