@@ -23,7 +23,7 @@ public class PostgreSqlTestContainer implements SqlTestContainer {
     @Override
     public void afterPropertiesSet() {
         if (null == postgreSQLContainer) {
-            postgreSQLContainer = new PostgreSQLContainer<>("postgres:16.2")
+            postgreSQLContainer = new PostgreSQLContainer<>("postgres:16.3")
                 .withDatabaseName("jhipsterSampleApplication")
                 .withTmpFs(Collections.singletonMap("/testtmpfs", "rw"))
                 .withLogConsumer(new Slf4jLogConsumer(log))
