@@ -20,6 +20,8 @@ import java.time.Instant;
 import java.util.*;
 import java.util.stream.Stream;
 import org.apache.commons.lang3.RandomStringUtils;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -57,6 +59,19 @@ class AccountResourceIT {
 
     @Autowired
     private MockMvc restAccountMockMvc;
+
+    private Long numberOfUsers;
+
+    @BeforeEach
+    public void countUsers() {
+        numberOfUsers = userRepository.count();
+    }
+
+    @AfterEach
+    public void cleanupAndCheck() {
+        assertThat(userRepository.count()).isEqualTo(numberOfUsers);
+        numberOfUsers = null;
+    }
 
     @Test
     @WithUnauthenticatedMockUser
@@ -103,6 +118,8 @@ class AccountResourceIT {
             .andExpect(jsonPath("$.imageUrl").value("http://placehold.it/50x50"))
             .andExpect(jsonPath("$.langKey").value("en"))
             .andExpect(jsonPath("$.authorities").value(AuthoritiesConstants.ADMIN));
+
+        userService.deleteUser(TEST_USER_LOGIN);
     }
 
     @Test
@@ -129,6 +146,8 @@ class AccountResourceIT {
             .andExpect(status().isCreated());
 
         assertThat(userRepository.findOneByLogin("test-register-valid")).isPresent();
+
+        userService.deleteUser("test-register-valid");
     }
 
     @Test
@@ -242,6 +261,8 @@ class AccountResourceIT {
         restAccountMockMvc
             .perform(post("/api/register").contentType(MediaType.APPLICATION_JSON).content(om.writeValueAsBytes(secondUser)))
             .andExpect(status().is4xxClientError());
+
+        userService.deleteUser("alice");
     }
 
     @Test
@@ -316,6 +337,8 @@ class AccountResourceIT {
         restAccountMockMvc
             .perform(post("/api/register").contentType(MediaType.APPLICATION_JSON).content(om.writeValueAsBytes(secondUser)))
             .andExpect(status().is4xxClientError());
+
+        userService.deleteUser("test-register-duplicate-email-3");
     }
 
     @Test
@@ -341,6 +364,8 @@ class AccountResourceIT {
         assertThat(userDup.orElseThrow().getAuthorities())
             .hasSize(1)
             .containsExactly(authorityRepository.findById(AuthoritiesConstants.USER).orElseThrow());
+
+        userService.deleteUser("badguy");
     }
 
     @Test
@@ -360,6 +385,8 @@ class AccountResourceIT {
 
         user = userRepository.findOneByLogin(user.getLogin()).orElse(null);
         assertThat(user.isActivated()).isTrue();
+
+        userService.deleteUser("activate-account");
     }
 
     @Test
@@ -402,6 +429,8 @@ class AccountResourceIT {
         assertThat(updatedUser.getImageUrl()).isEqualTo(userDTO.getImageUrl());
         assertThat(updatedUser.isActivated()).isTrue();
         assertThat(updatedUser.getAuthorities()).isEmpty();
+
+        userService.deleteUser("save-account");
     }
 
     @Test
@@ -431,6 +460,8 @@ class AccountResourceIT {
             .andExpect(status().isBadRequest());
 
         assertThat(userRepository.findOneByEmailIgnoreCase("invalid email")).isNotPresent();
+
+        userService.deleteUser("save-invalid-email");
     }
 
     @Test
@@ -466,8 +497,11 @@ class AccountResourceIT {
             .perform(post("/api/account").contentType(MediaType.APPLICATION_JSON).content(om.writeValueAsBytes(userDTO)))
             .andExpect(status().isBadRequest());
 
-        User updatedUser = userRepository.findOneByLogin("save-existing-email").orElse(null);
+        User updatedUser = userRepository.findOneByLogin("save-existing-email").orElseThrow();
         assertThat(updatedUser.getEmail()).isEqualTo("save-existing-email@example.com");
+
+        userService.deleteUser("save-existing-email");
+        userService.deleteUser("save-existing-email2");
     }
 
     @Test
@@ -497,6 +531,8 @@ class AccountResourceIT {
 
         User updatedUser = userRepository.findOneByLogin("save-existing-email-and-login").orElse(null);
         assertThat(updatedUser.getEmail()).isEqualTo("save-existing-email-and-login@example.com");
+
+        userService.deleteUser("save-existing-email-and-login");
     }
 
     @Test
@@ -521,6 +557,8 @@ class AccountResourceIT {
         User updatedUser = userRepository.findOneByLogin("change-password-wrong-existing-password").orElse(null);
         assertThat(passwordEncoder.matches("new password", updatedUser.getPassword())).isFalse();
         assertThat(passwordEncoder.matches(currentPassword, updatedUser.getPassword())).isTrue();
+
+        userService.deleteUser("change-password-wrong-existing-password");
     }
 
     @Test
@@ -544,6 +582,8 @@ class AccountResourceIT {
 
         User updatedUser = userRepository.findOneByLogin("change-password").orElse(null);
         assertThat(passwordEncoder.matches("new password", updatedUser.getPassword())).isTrue();
+
+        userService.deleteUser("change-password");
     }
 
     @Test
@@ -569,6 +609,8 @@ class AccountResourceIT {
 
         User updatedUser = userRepository.findOneByLogin("change-password-too-small").orElse(null);
         assertThat(updatedUser.getPassword()).isEqualTo(user.getPassword());
+
+        userService.deleteUser("change-password-too-small");
     }
 
     @Test
@@ -594,6 +636,8 @@ class AccountResourceIT {
 
         User updatedUser = userRepository.findOneByLogin("change-password-too-long").orElse(null);
         assertThat(updatedUser.getPassword()).isEqualTo(user.getPassword());
+
+        userService.deleteUser("change-password-too-long");
     }
 
     @Test
@@ -617,6 +661,8 @@ class AccountResourceIT {
 
         User updatedUser = userRepository.findOneByLogin("change-password-empty").orElse(null);
         assertThat(updatedUser.getPassword()).isEqualTo(user.getPassword());
+
+        userService.deleteUser("change-password-empty");
     }
 
     @Test
@@ -633,6 +679,8 @@ class AccountResourceIT {
         restAccountMockMvc
             .perform(post("/api/account/reset-password/init").content("password-reset@example.com"))
             .andExpect(status().isOk());
+
+        userService.deleteUser("password-reset");
     }
 
     @Test
@@ -649,6 +697,8 @@ class AccountResourceIT {
         restAccountMockMvc
             .perform(post("/api/account/reset-password/init").content("password-reset-upper-case@EXAMPLE.COM"))
             .andExpect(status().isOk());
+
+        userService.deleteUser("password-reset-upper-case");
     }
 
     @Test
@@ -683,6 +733,8 @@ class AccountResourceIT {
 
         User updatedUser = userRepository.findOneByLogin(user.getLogin()).orElse(null);
         assertThat(passwordEncoder.matches(keyAndPassword.getNewPassword(), updatedUser.getPassword())).isTrue();
+
+        userService.deleteUser("finish-password-reset");
     }
 
     @Test
@@ -710,6 +762,8 @@ class AccountResourceIT {
 
         User updatedUser = userRepository.findOneByLogin(user.getLogin()).orElse(null);
         assertThat(passwordEncoder.matches(keyAndPassword.getNewPassword(), updatedUser.getPassword())).isFalse();
+
+        userService.deleteUser("finish-password-reset-too-small");
     }
 
     @Test
