@@ -12,6 +12,7 @@ import io.github.jhipster.sample.domain.Authority;
 import io.github.jhipster.sample.repository.AuthorityRepository;
 import jakarta.persistence.EntityManager;
 import java.util.UUID;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -46,6 +47,8 @@ class AuthorityResourceIT {
 
     private Authority authority;
 
+    private Authority insertedAuthority;
+
     /**
      * Create an entity for this test.
      *
@@ -73,6 +76,14 @@ class AuthorityResourceIT {
         authority = createEntity(em);
     }
 
+    @AfterEach
+    public void cleanup() {
+        if (insertedAuthority != null) {
+            authorityRepository.delete(insertedAuthority);
+            insertedAuthority = null;
+        }
+    }
+
     @Test
     @Transactional
     void createAuthority() throws Exception {
@@ -91,13 +102,15 @@ class AuthorityResourceIT {
         // Validate the Authority in the database
         assertIncrementedRepositoryCount(databaseSizeBeforeCreate);
         assertAuthorityUpdatableFieldsEquals(returnedAuthority, getPersistedAuthority(returnedAuthority));
+
+        insertedAuthority = returnedAuthority;
     }
 
     @Test
     @Transactional
     void createAuthorityWithExistingId() throws Exception {
         // Create the Authority with an existing ID
-        authorityRepository.saveAndFlush(authority);
+        insertedAuthority = authorityRepository.saveAndFlush(authority);
 
         long databaseSizeBeforeCreate = getRepositoryCount();
 
@@ -115,7 +128,7 @@ class AuthorityResourceIT {
     void getAllAuthorities() throws Exception {
         // Initialize the database
         authority.setName(UUID.randomUUID().toString());
-        authorityRepository.saveAndFlush(authority);
+        insertedAuthority = authorityRepository.saveAndFlush(authority);
 
         // Get all the authorityList
         restAuthorityMockMvc
@@ -130,7 +143,7 @@ class AuthorityResourceIT {
     void getAuthority() throws Exception {
         // Initialize the database
         authority.setName(UUID.randomUUID().toString());
-        authorityRepository.saveAndFlush(authority);
+        insertedAuthority = authorityRepository.saveAndFlush(authority);
 
         // Get the authority
         restAuthorityMockMvc
@@ -152,7 +165,7 @@ class AuthorityResourceIT {
     void deleteAuthority() throws Exception {
         // Initialize the database
         authority.setName(UUID.randomUUID().toString());
-        authorityRepository.saveAndFlush(authority);
+        insertedAuthority = authorityRepository.saveAndFlush(authority);
 
         long databaseSizeBeforeDelete = getRepositoryCount();
 
