@@ -13,11 +13,13 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.github.jhipster.sample.IntegrationTest;
 import io.github.jhipster.sample.domain.BankAccount;
 import io.github.jhipster.sample.repository.BankAccountRepository;
+import io.github.jhipster.sample.repository.UserRepository;
 import jakarta.persistence.EntityManager;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Random;
 import java.util.concurrent.atomic.AtomicLong;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -59,6 +61,9 @@ class BankAccountResourceIT {
     @Autowired
     private BankAccountRepository bankAccountRepository;
 
+    @Autowired
+    private UserRepository userRepository;
+
     @Mock
     private BankAccountRepository bankAccountRepositoryMock;
 
@@ -69,6 +74,8 @@ class BankAccountResourceIT {
     private MockMvc restBankAccountMockMvc;
 
     private BankAccount bankAccount;
+
+    private BankAccount insertedBankAccount;
 
     /**
      * Create an entity for this test.
@@ -97,6 +104,14 @@ class BankAccountResourceIT {
         bankAccount = createEntity(em);
     }
 
+    @AfterEach
+    public void cleanup() {
+        if (insertedBankAccount != null) {
+            bankAccountRepository.delete(insertedBankAccount);
+            insertedBankAccount = null;
+        }
+    }
+
     @Test
     @Transactional
     void createBankAccount() throws Exception {
@@ -115,6 +130,8 @@ class BankAccountResourceIT {
         // Validate the BankAccount in the database
         assertIncrementedRepositoryCount(databaseSizeBeforeCreate);
         assertBankAccountUpdatableFieldsEquals(returnedBankAccount, getPersistedBankAccount(returnedBankAccount));
+
+        insertedBankAccount = returnedBankAccount;
     }
 
     @Test
@@ -170,7 +187,7 @@ class BankAccountResourceIT {
     @Transactional
     void getAllBankAccounts() throws Exception {
         // Initialize the database
-        bankAccountRepository.saveAndFlush(bankAccount);
+        insertedBankAccount = bankAccountRepository.saveAndFlush(bankAccount);
 
         // Get all the bankAccountList
         restBankAccountMockMvc
@@ -203,7 +220,7 @@ class BankAccountResourceIT {
     @Transactional
     void getBankAccount() throws Exception {
         // Initialize the database
-        bankAccountRepository.saveAndFlush(bankAccount);
+        insertedBankAccount = bankAccountRepository.saveAndFlush(bankAccount);
 
         // Get the bankAccount
         restBankAccountMockMvc
@@ -226,7 +243,7 @@ class BankAccountResourceIT {
     @Transactional
     void putExistingBankAccount() throws Exception {
         // Initialize the database
-        bankAccountRepository.saveAndFlush(bankAccount);
+        insertedBankAccount = bankAccountRepository.saveAndFlush(bankAccount);
 
         long databaseSizeBeforeUpdate = getRepositoryCount();
 
@@ -306,15 +323,13 @@ class BankAccountResourceIT {
     @Transactional
     void partialUpdateBankAccountWithPatch() throws Exception {
         // Initialize the database
-        bankAccountRepository.saveAndFlush(bankAccount);
+        insertedBankAccount = bankAccountRepository.saveAndFlush(bankAccount);
 
         long databaseSizeBeforeUpdate = getRepositoryCount();
 
         // Update the bankAccount using partial update
         BankAccount partialUpdatedBankAccount = new BankAccount();
         partialUpdatedBankAccount.setId(bankAccount.getId());
-
-        partialUpdatedBankAccount.name(UPDATED_NAME);
 
         restBankAccountMockMvc
             .perform(
@@ -337,7 +352,7 @@ class BankAccountResourceIT {
     @Transactional
     void fullUpdateBankAccountWithPatch() throws Exception {
         // Initialize the database
-        bankAccountRepository.saveAndFlush(bankAccount);
+        insertedBankAccount = bankAccountRepository.saveAndFlush(bankAccount);
 
         long databaseSizeBeforeUpdate = getRepositoryCount();
 
@@ -418,7 +433,7 @@ class BankAccountResourceIT {
     @Transactional
     void deleteBankAccount() throws Exception {
         // Initialize the database
-        bankAccountRepository.saveAndFlush(bankAccount);
+        insertedBankAccount = bankAccountRepository.saveAndFlush(bankAccount);
 
         long databaseSizeBeforeDelete = getRepositoryCount();
 
