@@ -14,6 +14,7 @@ import io.github.jhipster.sample.repository.LabelRepository;
 import jakarta.persistence.EntityManager;
 import java.util.Random;
 import java.util.concurrent.atomic.AtomicLong;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -54,6 +55,8 @@ class LabelResourceIT {
 
     private Label label;
 
+    private Label insertedLabel;
+
     /**
      * Create an entity for this test.
      *
@@ -81,6 +84,14 @@ class LabelResourceIT {
         label = createEntity(em);
     }
 
+    @AfterEach
+    public void cleanup() {
+        if (insertedLabel != null) {
+            labelRepository.delete(insertedLabel);
+            insertedLabel = null;
+        }
+    }
+
     @Test
     @Transactional
     void createLabel() throws Exception {
@@ -99,6 +110,8 @@ class LabelResourceIT {
         // Validate the Label in the database
         assertIncrementedRepositoryCount(databaseSizeBeforeCreate);
         assertLabelUpdatableFieldsEquals(returnedLabel, getPersistedLabel(returnedLabel));
+
+        insertedLabel = returnedLabel;
     }
 
     @Test
@@ -138,7 +151,7 @@ class LabelResourceIT {
     @Transactional
     void getAllLabels() throws Exception {
         // Initialize the database
-        labelRepository.saveAndFlush(label);
+        insertedLabel = labelRepository.saveAndFlush(label);
 
         // Get all the labelList
         restLabelMockMvc
@@ -153,7 +166,7 @@ class LabelResourceIT {
     @Transactional
     void getLabel() throws Exception {
         // Initialize the database
-        labelRepository.saveAndFlush(label);
+        insertedLabel = labelRepository.saveAndFlush(label);
 
         // Get the label
         restLabelMockMvc
@@ -175,7 +188,7 @@ class LabelResourceIT {
     @Transactional
     void putExistingLabel() throws Exception {
         // Initialize the database
-        labelRepository.saveAndFlush(label);
+        insertedLabel = labelRepository.saveAndFlush(label);
 
         long databaseSizeBeforeUpdate = getRepositoryCount();
 
@@ -251,7 +264,7 @@ class LabelResourceIT {
     @Transactional
     void partialUpdateLabelWithPatch() throws Exception {
         // Initialize the database
-        labelRepository.saveAndFlush(label);
+        insertedLabel = labelRepository.saveAndFlush(label);
 
         long databaseSizeBeforeUpdate = getRepositoryCount();
 
@@ -277,7 +290,7 @@ class LabelResourceIT {
     @Transactional
     void fullUpdateLabelWithPatch() throws Exception {
         // Initialize the database
-        labelRepository.saveAndFlush(label);
+        insertedLabel = labelRepository.saveAndFlush(label);
 
         long databaseSizeBeforeUpdate = getRepositoryCount();
 
@@ -356,7 +369,7 @@ class LabelResourceIT {
     @Transactional
     void deleteLabel() throws Exception {
         // Initialize the database
-        labelRepository.saveAndFlush(label);
+        insertedLabel = labelRepository.saveAndFlush(label);
 
         long databaseSizeBeforeDelete = getRepositoryCount();
 
