@@ -20,6 +20,7 @@ import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Random;
 import java.util.concurrent.atomic.AtomicLong;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -75,6 +76,8 @@ class OperationResourceIT {
 
     private Operation operation;
 
+    private Operation insertedOperation;
+
     /**
      * Create an entity for this test.
      *
@@ -102,6 +105,14 @@ class OperationResourceIT {
         operation = createEntity(em);
     }
 
+    @AfterEach
+    public void cleanup() {
+        if (insertedOperation != null) {
+            operationRepository.delete(insertedOperation);
+            insertedOperation = null;
+        }
+    }
+
     @Test
     @Transactional
     void createOperation() throws Exception {
@@ -120,6 +131,8 @@ class OperationResourceIT {
         // Validate the Operation in the database
         assertIncrementedRepositoryCount(databaseSizeBeforeCreate);
         assertOperationUpdatableFieldsEquals(returnedOperation, getPersistedOperation(returnedOperation));
+
+        insertedOperation = returnedOperation;
     }
 
     @Test
@@ -175,7 +188,7 @@ class OperationResourceIT {
     @Transactional
     void getAllOperations() throws Exception {
         // Initialize the database
-        operationRepository.saveAndFlush(operation);
+        insertedOperation = operationRepository.saveAndFlush(operation);
 
         // Get all the operationList
         restOperationMockMvc
@@ -209,7 +222,7 @@ class OperationResourceIT {
     @Transactional
     void getOperation() throws Exception {
         // Initialize the database
-        operationRepository.saveAndFlush(operation);
+        insertedOperation = operationRepository.saveAndFlush(operation);
 
         // Get the operation
         restOperationMockMvc
@@ -233,7 +246,7 @@ class OperationResourceIT {
     @Transactional
     void putExistingOperation() throws Exception {
         // Initialize the database
-        operationRepository.saveAndFlush(operation);
+        insertedOperation = operationRepository.saveAndFlush(operation);
 
         long databaseSizeBeforeUpdate = getRepositoryCount();
 
@@ -311,7 +324,7 @@ class OperationResourceIT {
     @Transactional
     void partialUpdateOperationWithPatch() throws Exception {
         // Initialize the database
-        operationRepository.saveAndFlush(operation);
+        insertedOperation = operationRepository.saveAndFlush(operation);
 
         long databaseSizeBeforeUpdate = getRepositoryCount();
 
@@ -319,7 +332,7 @@ class OperationResourceIT {
         Operation partialUpdatedOperation = new Operation();
         partialUpdatedOperation.setId(operation.getId());
 
-        partialUpdatedOperation.description(UPDATED_DESCRIPTION);
+        partialUpdatedOperation.date(UPDATED_DATE);
 
         restOperationMockMvc
             .perform(
@@ -342,7 +355,7 @@ class OperationResourceIT {
     @Transactional
     void fullUpdateOperationWithPatch() throws Exception {
         // Initialize the database
-        operationRepository.saveAndFlush(operation);
+        insertedOperation = operationRepository.saveAndFlush(operation);
 
         long databaseSizeBeforeUpdate = getRepositoryCount();
 
@@ -423,7 +436,7 @@ class OperationResourceIT {
     @Transactional
     void deleteOperation() throws Exception {
         // Initialize the database
-        operationRepository.saveAndFlush(operation);
+        insertedOperation = operationRepository.saveAndFlush(operation);
 
         long databaseSizeBeforeDelete = getRepositoryCount();
 
