@@ -31,7 +31,7 @@ describe('login modal', () => {
   it('requires username', () => {
     cy.get(passwordLoginSelector).type('a-password');
     cy.get(submitLoginSelector).click();
-    cy.wait('@authenticate').then(({ response }) => expect(response.statusCode).to.equal(400));
+    cy.wait('@authenticate').then(({ response }) => expect(response?.statusCode).to.equal(400));
     // login page should stay open when login fails
     cy.get(titleLoginSelector).should('be.visible');
   });
@@ -39,7 +39,7 @@ describe('login modal', () => {
   it('requires password', () => {
     cy.get(usernameLoginSelector).type('a-login');
     cy.get(submitLoginSelector).click();
-    cy.wait('@authenticate').then(({ response }) => expect(response.statusCode).to.equal(400));
+    cy.wait('@authenticate').then(({ response }) => expect(response?.statusCode).to.equal(400));
     cy.get(errorLoginSelector).should('be.visible');
   });
 
@@ -47,7 +47,7 @@ describe('login modal', () => {
     cy.get(usernameLoginSelector).type(username);
     cy.get(passwordLoginSelector).type('bad-password');
     cy.get(submitLoginSelector).click();
-    cy.wait('@authenticate').then(({ response }) => expect(response.statusCode).to.equal(401));
+    cy.wait('@authenticate').then(({ response }) => expect(response?.statusCode).to.equal(401));
     cy.get(errorLoginSelector).should('be.visible');
   });
 
@@ -55,7 +55,7 @@ describe('login modal', () => {
     cy.get(usernameLoginSelector).type(username);
     cy.get(passwordLoginSelector).type(password);
     cy.get(submitLoginSelector).click();
-    cy.wait('@authenticate').then(({ response }) => expect(response.statusCode).to.equal(200));
+    cy.wait('@authenticate').then(({ response }) => expect(response?.statusCode).to.equal(200));
     cy.hash().should('eq', '');
   });
 });
