@@ -30,6 +30,6 @@ describe('forgot your password', () => {
   it('should be able to init reset password', () => {
     cy.get(emailResetPasswordSelector).type('user@gmail.com');
     cy.get(submitInitResetPasswordSelector).click({ force: true });
-    cy.wait('@initResetPassword').then(({ response }) => expect(response.statusCode).to.equal(200));
+    cy.wait('@initResetPassword').then(({ response }) => expect(response?.statusCode).to.equal(200));
   });
 });
