@@ -15,7 +15,7 @@ describe('Operation e2e test', () => {
   const operationPageUrlPattern = new RegExp('/operation(\\?.*)?$');
   const username = Cypress.env('E2E_USERNAME') ?? 'user';
   const password = Cypress.env('E2E_PASSWORD') ?? 'user';
-  const operationSample = { date: '2015-08-05T04:20:57.915Z', amount: 21503.17 };
+  const operationSample = { date: '2015-08-05T08:14:11.129Z', amount: 16346.49 };
 
   let operation;
 
@@ -44,7 +44,7 @@ describe('Operation e2e test', () => {
     cy.visit('/');
     cy.clickOnEntityMenuItem('operation');
     cy.wait('@entitiesRequest').then(({ response }) => {
-      if (response.body.length === 0) {
+      if (response?.body.length === 0) {
         cy.get(entityTableSelector).should('not.exist');
       } else {
         cy.get(entityTableSelector).should('exist');
@@ -68,7 +68,7 @@ describe('Operation e2e test', () => {
         cy.get(entityCreateSaveButtonSelector).should('exist');
         cy.get(entityCreateCancelButtonSelector).click();
         cy.wait('@entitiesRequest').then(({ response }) => {
-          expect(response.statusCode).to.equal(200);
+          expect(response?.statusCode).to.equal(200);
         });
         cy.url().should('match', operationPageUrlPattern);
       });
@@ -109,7 +109,7 @@ describe('Operation e2e test', () => {
         cy.getEntityDetailsHeading('operation');
         cy.get(entityDetailsBackButtonSelector).click();
         cy.wait('@entitiesRequest').then(({ response }) => {
-          expect(response.statusCode).to.equal(200);
+          expect(response?.statusCode).to.equal(200);
         });
         cy.url().should('match', operationPageUrlPattern);
       });
@@ -120,7 +120,7 @@ describe('Operation e2e test', () => {
         cy.get(entityCreateSaveButtonSelector).should('exist');
         cy.get(entityCreateCancelButtonSelector).click();
         cy.wait('@entitiesRequest').then(({ response }) => {
-          expect(response.statusCode).to.equal(200);
+          expect(response?.statusCode).to.equal(200);
         });
         cy.url().should('match', operationPageUrlPattern);
       });
@@ -130,7 +130,7 @@ describe('Operation e2e test', () => {
         cy.getEntityCreateUpdateHeading('Operation');
         cy.get(entityCreateSaveButtonSelector).click();
         cy.wait('@entitiesRequest').then(({ response }) => {
-          expect(response.statusCode).to.equal(200);
+          expect(response?.statusCode).to.equal(200);
         });
         cy.url().should('match', operationPageUrlPattern);
       });
@@ -140,10 +140,10 @@ describe('Operation e2e test', () => {
         cy.getEntityDeleteDialogHeading('operation').should('exist');
         cy.get(entityConfirmDeleteButtonSelector).click();
         cy.wait('@deleteEntityRequest').then(({ response }) => {
-          expect(response.statusCode).to.equal(204);
+          expect(response?.statusCode).to.equal(204);
         });
         cy.wait('@entitiesRequest').then(({ response }) => {
-          expect(response.statusCode).to.equal(200);
+          expect(response?.statusCode).to.equal(200);
         });
         cy.url().should('match', operationPageUrlPattern);
 
@@ -160,24 +160,24 @@ describe('Operation e2e test', () => {
     });
 
     it('should create an instance of Operation', () => {
-      cy.get(`[data-cy="date"]`).type('2015-08-05T09:37');
+      cy.get(`[data-cy="date"]`).type('2015-08-05T01:32');
       cy.get(`[data-cy="date"]`).blur();
-      cy.get(`[data-cy="date"]`).should('have.value', '2015-08-05T09:37');
+      cy.get(`[data-cy="date"]`).should('have.value', '2015-08-05T01:32');
 
-      cy.get(`[data-cy="description"]`).type('oof');
-      cy.get(`[data-cy="description"]`).should('have.value', 'oof');
+      cy.get(`[data-cy="description"]`).type('dense');
+      cy.get(`[data-cy="description"]`).should('have.value', 'dense');
 
-      cy.get(`[data-cy="amount"]`).type('27579.12');
-      cy.get(`[data-cy="amount"]`).should('have.value', '27579.12');
+      cy.get(`[data-cy="amount"]`).type('28492.17');
+      cy.get(`[data-cy="amount"]`).should('have.value', '28492.17');
 
       cy.get(entityCreateSaveButtonSelector).click();
 
       cy.wait('@postEntityRequest').then(({ response }) => {
-        expect(response.statusCode).to.equal(201);
+        expect(response?.statusCode).to.equal(201);
         operation = response.body;
       });
       cy.wait('@entitiesRequest').then(({ response }) => {
-        expect(response.statusCode).to.equal(200);
+        expect(response?.statusCode).to.equal(200);
       });
       cy.url().should('match', operationPageUrlPattern);
     });
